Déplacement de l'affichage de l'utilisation du disque vers la classe MemoryMonitor
